change mapper for import and export
It now looks recursively through the given path. Each time it finds a folder called 'mapping' it will collect the robots inside that folder.
This way import AND export templates will be found. A '<connectorName>Mapper.xill' is made for each Connector it finds.
These new robots are placed in the UDM folder of the connectors